End day @home 11-03-2021

+++ groceries_php/database/migrations/2021_03_10_125837_create_products_table.php
@@ -18,13 +18,9 @@ class CreateProductsTable extends Migration
             $table->text('product_name');
             $table->text('product_description');
             $table->decimal('product_price', $precision = 8, $scale = 2);
-            $table->int('product_amount');
+            $table->bigInteger('product_amount');
             $table->timestamps();
         });
-
-
-
-
     }
 
     /**
